Ba feature/routes

+++ controllers/api/comment-routes.js
@@ -2,15 +2,18 @@ const router = require('express').Router();
 const { Comment } = require('../../models');
 
 router.get('/', (req, res) => {
+  if (req.session) {
     Comment.findAll()
     .then(dbCommentData => res.json(dbCommentData))
     .catch(err => {
       console.log(err);
       res.status(500).json(err);
     });
+  }
 });
 
 router.post('/', (req, res) => {
+  if (req.session) {
     Comment.create({
         comment_text: req.body.comment_text,
         user_id: req.body.user_id,
@@ -21,10 +24,11 @@ router.post('/', (req, res) => {
           console.log(err);
           res.status(400).json(err);
         });
-
+  }
 });
 
 router.delete('/:id', (req, res) => {
+  if (req.session) {
     Comment.destroy({
         where: {
           id: req.params.id
@@ -41,6 +45,7 @@ router.delete('/:id', (req, res) => {
           console.log(err);
           res.status(500).json(err);
         });
+  }
 });
 
 module.exports = router;
